+++ CHANGELOG.md
@@ -6,11 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+
+
+## [0.2.1] - 2017-09-27
+### Added
 - `install:git` command
 - `delete:git` command
 - `install:wordpress` command
 - `delete:wordpress` command
 - Ability to inline file contents to `make:recipe`, `env:update`, etc.
+- Started adding tests
 
 ## [0.2.0] - 2017-06-23
 ### Added
@@ -113,7 +118,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `update:site` command
 - `delete:site` command
 
-[Unreleased]: https://github.com/svenluijten/forge-cli/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/svenluijten/forge-cli/compare/0.2.1...HEAD
+[0.2.1]: https://github.com/svenluijten/forge-cli/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/svenluijten/forge-cli/compare/0.1.5...0.2.0
 [0.1.5]: https://github.com/svenluijten/forge-cli/compare/0.1.4...0.1.5
 [0.1.4]: https://github.com/svenluijten/forge-cli/compare/0.1.3...0.1.4